+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Mulberry_Warranty changelog
 ========================
 
+1.5.0:
+- Added force logging functionality for the Mulberry API calls
+- Added separate CLI command to be able to force sync order & cart hooks using the Magento order_id param
+
 1.4.1:
 - Added CSP whitelist records for Mulberry API